% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGBIF.R
\name{clean_gbif}
\alias{clean_gbif}
\title{Clean GBIF data}
\usage{
clean_gbif(
  raw,
  startYear = 1900,
  endYear = as.numeric(format(Sys.Date(), "\%Y")),
  centroidBufferKm = 2000,
  capitalBufferKm = 2000,
  institutionBufferKm = 2000
)
}
\arguments{
\item{raw}{(data.frame) Output from download_gbif()}

\item{startYear}{(numeric) Start year of data}

\item{endYear}{(numeric) End year of data}

\item{centroidBufferKm}{(numeric) Threshold for distance from state/county/etc centroid}

\item{capitalBufferKm}{(numeric) Threshold for distance from capitol}

\item{institutionBufferKm}{(numeric) Threshold for distance from institution (zoo, arboretum, etc.)}
}
\value{
A cleaned GBIF dataset
}
\description{
Clean raw GBIF data. Removes duplicates and coordinates which are 2km within a country centroid, a city, or a zoo/herbarium.
}
\examples{
\dontrun{
raw <- download_gbif(scientificName = "Anaxyrus macroscaphus",
                     user = "Username",
                     pwd = "Password",
                     email = "email@email.com")

clean <- clean_gbif(raw = raw$dat,
                    startYear = 1980)
}
}
