% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGBIF.R
\name{download_gbif}
\alias{download_gbif}
\title{Download GBIF data}
\usage{
download_gbif(scientificName, user, pwd, email)
}
\arguments{
\item{scientificName}{(character) Scientific name}

\item{user}{(character) Username for iNaturalist account}

\item{pwd}{(character) Password for iNaturalist account}

\item{email}{(character) Email address for iNaturalist account}
}
\description{
Download GBIF data
}
