% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGBIF.R
\name{process_gbif}
\alias{process_gbif}
\title{processGBIF}
\usage{
process_gbif(
  scientificName,
  keep,
  sp.code,
  user,
  pwd,
  email,
  startYear = 1900,
  endYear = as.numeric(format(Sys.Date(), "\%Y")),
  coordPrec = 0.01,
  centroidBufferKm = 2000,
  capitalBufferKm = 2000,
  institutionBufferKm = 2000
)
}
\arguments{
\item{institutionBufferKm}{}
}
\description{
processGBIF
}
