% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGBIF.R
\name{clean_gbif}
\alias{clean_gbif}
\title{Title}
\usage{
clean_gbif(
  raw,
  startYear = 1900,
  endYear = as.numeric(format(Sys.Date(), "\%Y")),
  coordPrec = 0.01,
  centroidBufferKm = 2000,
  capitalBufferKm = 2000,
  institutionBufferKm = 2000
)
}
\arguments{
\item{institutionBufferKm}{}
}
\description{
Title
}
