% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGBIF.R
\name{download_gbif}
\alias{download_gbif}
\title{Download GBIF data}
\usage{
download_gbif(
  scientificName,
  country = "US",
  source = "all",
  startYear = 1900,
  user,
  pwd,
  email,
  taxonKey = "unknown"
)
}
\arguments{
\item{scientificName}{(character) Scientific name}

\item{country}{(character) Country to download data from}

\item{source}{(character) Source to download data from (e.g. "iNaturalist")}

\item{startYear}{(numeric) Year after which to download data}

\item{user}{(character) Username for GBIF account}

\item{pwd}{(character) Password for GBIF account}

\item{email}{(character) Email address for GBIF account}

\item{taxonKey}{(numeric) Taxon key representing the species of interest in GBIF. Default is "unknown", meaning taxon key will be searched for based on Scientific Name}
}
\value{
A list of GBIF data and the associated citation
}
\description{
Download GBIF data
}
\examples{
\dontrun{
raw <- download_gbif(scientificName = "Anaxyrus macroscaphus",
                     country = "US",
                     source = "all",
                     user = "Username",
                     pwd = "Password",
                     email = "email@email.com",
                     taxonKey = "unknown")
}
}
