% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGBIF.R
\name{download_gbif}
\alias{download_gbif}
\title{Download GBIF data}
\usage{
download_gbif(
  scientificName,
  startyear,
  country = "US",
  source = "all",
  user,
  pwd,
  email
)
}
\arguments{
\item{scientificName}{(character) Scientific name}

\item{startyear}{(numeric) First year to download data from}

\item{country}{(character) Country to download data from}

\item{source}{(character) Source to download data from (e.g. "iNaturalist")}

\item{user}{(character) Username for GBIF account}

\item{pwd}{(character) Password for GBIF account}

\item{email}{(character) Email address for GBIF account}
}
\value{
A list of GBIF data and the associated citation
}
\description{
Download GBIF data
}
\examples{
\dontrun{
raw <- download_gbif(scientificName = "Anaxyrus macroscaphus",
                     startyear = 1990,
                     country = "US",
                     source = "all",
                     user = "Username",
                     pwd = "Password",
                     email = "email@email.com")
}
}
