% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGBIF.R
\name{process_gbif}
\alias{process_gbif}
\title{Download and Process GBIF Data}
\usage{
process_gbif(
  scientificName,
  keep,
  sp.code,
  user,
  pwd,
  email,
  startYear = 1900,
  endYear = as.numeric(format(Sys.Date(), "\%Y")),
  centroidBufferKm = 2000,
  capitalBufferKm = 2000,
  institutionBufferKm = 2000
)
}
\arguments{
\item{scientificName}{(character) Scientific name}

\item{keep}{(character) Must be "iNaturalist" or "Museum"; Not case sensitive}

\item{sp.code}{(character) User-provided four character species code}

\item{user}{(character) Username for GBIF account}

\item{pwd}{(character) Password for GBIF account}

\item{email}{(character) Email address for GBIF account}

\item{startYear}{(numeric) Start year of data}

\item{endYear}{(numeric) End year of data}

\item{centroidBufferKm}{(numeric) Threshold for distance from state/county/etc centroid}

\item{capitalBufferKm}{(numeric) Threshold for distance from capitol}

\item{institutionBufferKm}{(numeric) Threshold for distance from institution (zoo, arboretum, etc.)}
}
\value{
A list of GBIF data and the associated citation
}
\description{
A wrapper for download_gbif() and clean_gbif() to process data in one step
}
\examples{
\dontrun{
dat <- process_gbif(scientificName = "Anaxyrus microscaphus",
                    sp.code = "ANMI",
                    keep = "iNat",
                    user = "Username",
                    pwd = "Password",
                    email = "email@email.com")
}
}
