% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGBIF.R
\name{process_gbif}
\alias{process_gbif}
\title{Download iNaturalist and museum data stored in GBIF and format it for flexiSDM analysis}
\usage{
process_gbif(
  scientificName,
  sp.code = scientificName,
  keep = c("iNat", "museum"),
  data.path = ".",
  citation.path = ".",
  user,
  pwd,
  email,
  startYear = 1900,
  endYear = as.numeric(format(Sys.Date(), "\%Y")),
  centroidBufferKm = 2000,
  capitalBufferKm = 2000,
  institutionBufferKm = 2000
)
}
\arguments{
\item{scientificName}{(character vector) A vector of scientific names to search for}

\item{sp.code}{(character) User-provided species identifier; defaults to scientificName}

\item{keep}{(character vector) Vector including 'iNat' and/or 'museum'; default is both}

\item{data.path}{(character) Location where data should be saved; default is current directory}

\item{citation.path}{(character) Location where citation should be saved; default is current directory}

\item{user}{(character) Username for GBIF account}

\item{pwd}{(character) Password for GBIF account}

\item{email}{(character) Email address for GBIF account}

\item{startYear}{(numeric) Start year of data}

\item{endYear}{(numeric) End year of data}

\item{centroidBufferKm}{(numeric) Threshold for distance from state/county/etc centroid}

\item{capitalBufferKm}{(numeric) Threshold for distance from capitol}

\item{institutionBufferKm}{(numeric) Threshold for distance from institution (zoo, arboretum, etc.)}
}
\value{
Writes two csv files containing iNaturalist and museum data from GBIF, and two txt files containing the associated citation(s)
}
\description{
A wrapper for download_gbif() and clean_gbif() to process and format data for flexiSDM
}
\examples{
\dontrun{
dat <- process_gbif(scientificName = "Anaxyrus microscaphus",
                    sp.code = "ANMI",
                    user = "Username",
                    pwd = "Password",
                    email = "email@email.com")
}
}
